Planet Gravity Sets
Planet's gravity only activates when they're part of the current
gravity challenge between two checkpoints. This involves giving planets
different numbers based on which checkpoint they are related to. In the
future I may want to look into making the previous and current planets
gravity active, but I'm not sure.